aux: add separate components folder
