+++ solara/envs/components/battery.py
