+++ solara/envs/components/solar.py
